Поиск описания переменных модуля, обобщено в Trees
