+++ src/test/resources/providers/documentSymbol.bsl
@@ -1,6 +1,6 @@
 #Область Объявление
-
-Перем А1;
+// Это экспортная переменная
+Перем А1 Экспорт;
 
 Перем Б1, В1;
